requirements.txt
apify-client>=1.7
openai>=1.30
pandas>=2.0
streamlit>=1.35
python-dotenv>=1.0
requests>=2.31
pytest>=8.0
psycopg2-binary>=2.9
testing.postgresql>=1.3
